+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [3.4.1](https://github.com/xability/maidr/compare/v3.4.0...v3.4.1) (2025-05-25)
+
+
+### Bug Fixes
+
+* assign application role to svg upon activation ([#277](https://github.com/xability/maidr/issues/277)) ([e0a5a43](https://github.com/xability/maidr/commit/e0a5a43b03036c283276ef1659a0d1b10e5a31d4))
+
+# [3.4.0](https://github.com/xability/maidr/compare/v3.3.2...v3.4.0) (2025-05-23)
+
+
+### Features
+
+* support smooth continuous audio ([#268](https://github.com/xability/maidr/issues/268)) ([7ef714d](https://github.com/xability/maidr/commit/7ef714d5ff6fe0b4a97791696b5f5f4276e992cb))
+
 ## [3.3.2](https://github.com/xability/maidr/compare/v3.3.1...v3.3.2) (2025-05-20)